+++ src/main/java/http/Utils.java
@@ -1,4 +1,4 @@
-package util;
+package http;
 
 import java.nio.file.Path;
